Actualizacion
Se realizaron modificaciones en el Frontend

+++ Proyecto_Ciclo_3/Ciclo3demo/src/main/java/co/edu/unbosque/model/DetalleVenta.java
@@ -4,11 +4,13 @@ public class DetalleVenta {
 	
 	private Long codigo_detalle_venta;
 	private Integer cantidad_producto;
-	private Integer codigo_producto;
+	private Long codigo_producto;
 	private Long codigo_venta;
-	private Long valor_total;
-	private Long valor_venta;
-	private Long valor_iva;
+	private double valor_total;
+	private double valor_venta;
+	private double valor_iva;
+	
+	
 	public Long getCodigo_detalle_venta() {
 		return codigo_detalle_venta;
 	}
@@ -21,10 +23,10 @@ public class DetalleVenta {
 	public void setCantidad_producto(Integer cantidad_producto) {
 		this.cantidad_producto = cantidad_producto;
 	}
-	public Integer getCodigo_producto() {
+	public Long getCodigo_producto() {
 		return codigo_producto;
 	}
-	public void setCodigo_producto(Integer codigo_producto) {
+	public void setCodigo_producto(Long codigo_producto) {
 		this.codigo_producto = codigo_producto;
 	}
 	public Long getCodigo_venta() {
@@ -33,25 +35,23 @@ public class DetalleVenta {
 	public void setCodigo_venta(Long codigo_venta) {
 		this.codigo_venta = codigo_venta;
 	}
-	public Long getValor_total() {
+	public double getValor_total() {
 		return valor_total;
 	}
-	public void setValor_total(Long valor_total) {
+	public void setValor_total(double valor_total) {
 		this.valor_total = valor_total;
 	}
-	public Long getValor_venta() {
+	public double getValor_venta() {
 		return valor_venta;
 	}
-	public void setValor_venta(Long valor_venta) {
+	public void setValor_venta(double valor_venta) {
 		this.valor_venta = valor_venta;
 	}
-	public Long getValor_iva() {
+	public double getValor_iva() {
 		return valor_iva;
 	}
-	public void setValor_iva(Long valor_iva) {
+	public void setValor_iva(double valor_iva) {
 		this.valor_iva = valor_iva;
 	}
 	
-	
-
 }
