+++ Proyecto_Ciclo_3/ciclo3back/src/main/java/co/edu/unbosque/ciclo3back/model/Detalle_Venta.java
@@ -10,13 +10,13 @@ public class Detalle_Venta {
 	
 	@Id
 	@GeneratedValue (strategy=GenerationType.IDENTITY)//es el id y se va a generar de manera automatica e incremental, es lo que se dice con este @
-	private Integer codigo_detalle_venta;
+	private Long codigo_detalle_venta;
 	private Integer cantidad_producto;
 	private Long codigo_producto;
 	private Long codigo_venta;
-	private Long valor_total;
-	private Long valor_venta;
-	private Long valor_iva;	
+	private double valor_total;
+	private double valor_venta;
+	private double valor_iva;	
 	
 
 }
